fix: compile errors on windows-sys 0.52 (HWND is isize, Gdi feature, Timelike, winreg &str)
- windows-sys 0.52: handle types (HWND/HINSTANCE/HMENU/...) are isize,
  pass 0 for null handles instead of null_mut(); lpParam is *const c_void
- add Win32_Graphics_Gdi (WNDCLASSW/RegisterClassW) and Win32_System_Console (GetConsoleWindow) features
- UINT does not exist in windows-sys 0.52, use u32 for message type
- SendMessageTimeoutW lpdwresult is *mut usize
- chrono: hour()/minute()/second() come from Timelike, not Datelike
- winreg 0.52 set_value takes &T, string must be passed as &&str

+++ src/theme.rs
@@ -36,7 +36,7 @@ pub fn set_theme(theme: Theme) -> io::Result<()> {
             param.as_ptr() as LPARAM,
             SMTO_ABORTIFHUNG,
             100,
-            std::ptr::null_mut::<u32>(),
+            std::ptr::null_mut::<usize>(), // lpdwresult 是 *mut usize
         );
     }
     Ok(())
